Move BinarySerializableSerializer to the correct namespace
