+++ SAGESharp/IO/BinarySerializers.cs
@@ -285,34 +285,5 @@ namespace SAGESharp.IO
             binaryWriter.WriteUInt32(FOOTER_MAGIC_NUMBER);
         }
     }
-
-    internal sealed class BinarySerializableSerializer<T> : IBinarySerializer<T> where T : IBinarySerializable, new()
-    {
-        private readonly Func<T> constructor;
-
-        public BinarySerializableSerializer()
-        {
-            constructor = () => (T)Activator.CreateInstance(typeof(T), null);
-        }
-
-        public T Read(IBinaryReader binaryReader)
-        {
-            Validate.ArgumentNotNull(binaryReader, nameof(binaryReader));
-
-            T result = constructor();
-
-            result.Read(binaryReader);
-
-            return result;
-        }
-
-        public void Write(IBinaryWriter binaryWriter, T value)
-        {
-            Validate.ArgumentNotNull(binaryWriter, nameof(binaryWriter));
-            Validate.ArgumentNotNull<object>(value, nameof(value));
-
-            value.Write(binaryWriter);
-        }
-    }
     #endregion
 }
